applied background to all pages and added formatting
